Refactored file structure for premade services.
