+++ net/nimble/host/services/gap/include/services/gap/ble_svc_gap.h
